+++ spotbugs/src/main/java/edu/umd/cs/findbugs/workflow/Update.java
@@ -185,12 +185,7 @@ public class Update {
         matchBugs(baselineCollection, bugCollection);
         matchBugs(SortedBugCollection.BugInstanceComparator.instance, baselineCollection, bugCollection);
         matchBugs(versionInsensitiveBugComparator, baselineCollection, bugCollection);
-        for (Iterator<BugInstance> i = bugCollection.getCollection().iterator(); i.hasNext();) {
-            BugInstance bug = i.next();
-            if (matchedOldBugs.containsKey(bug)) {
-                i.remove();
-            }
-        }
+        bugCollection.getCollection().removeIf(matchedOldBugs::containsKey);
 
     }